Renamed, and Organized the file structure, Organized all imports and added services folder for all common api requests
