+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.6.5",
     "bootstrap": "^5.3.2",
     "leaflet": "^1.9.4",
     "react": "^18.2.0",
