+++ src/App scratch.js
@@ -1,13 +1,16 @@
 import './App.css';
+
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Homepage from './Pages/Homepage';
-import Login from './Pages/Login/Login';
-import HomepageStatic from './Pages/HomepageStatic';
+
+import Homepage from './pages/Homepage';
+import Login from './pages/login/Login';
+import HomepageStatic from './pages/HomepageStatic';
+
 function App() {
   return (
     <>
-<Homepage></Homepage>
+      <Homepage></Homepage>
     </>
   );
 }
