+++ src/assets/drawericon.png
