+++ src/assets/navicon.png
